feat:  ac vue-app in vue-3
