+++ vue-app/vite-vue-ts/package.json
@@ -10,7 +10,8 @@
   },
   "dependencies": {
     "pinia": "2.2.6",
-    "vue": "3.5.17"
+    "vue": "3.5.17",
+    "vue-router": "^4.5.1"
   },
   "devDependencies": {
     "@vitejs/plugin-vue": "6.0.0",
