+++ vue-app/vite-vue-ts/src/router/index.ts
@@ -1,10 +1,10 @@
-import { createRouter, createWebHistory, RouteRecordRaw } from 'vue-router'
+import { createRouter, createWebHistory, type RouteRecordRaw } from 'vue-router'
 
 const routes: RouteRecordRaw[] = [
   {
     path: '/',
     name: 'Home',
-    component: () => import('../views/Home.vue')
+    component: () => import('../views/Home/index.vue')
   },
   {
     path: '/vue3',
@@ -14,29 +14,29 @@ const routes: RouteRecordRaw[] = [
       {
         path: 'composition-api',
         name: 'CompositionAPI',
-        component: () => import('../views/vue3/CompositionAPI.vue')
+        component: () => import('../views/vue3/CompositionAPI/index.vue')
       },
       {
         path: 'reactive-system',
         name: 'ReactiveSystem',
-        component: () => import('../views/vue3/ReactiveSystem.vue')
+        component: () => import('../views/vue3/ReactiveSystem/index.vue')
       },
       {
         path: 'lifecycle-hooks',
         name: 'LifecycleHooks',
-        component: () => import('../views/vue3/LifecycleHooks.vue')
+        component: () => import('../views/vue3/LifecycleHooks/index.vue')
       },
       {
         path: 'performance',
         name: 'Performance',
-        component: () => import('../views/vue3/Performance.vue')
+        component: () => import('../views/vue3/Performance/index.vue')
       }
     ]
   },
   {
     path: '/api-compare',
     name: 'APICompare',
-    component: () => import('../views/APICompare.vue')
+    component: () => import('../views/APICompare/index.vue')
   },
   {
     path: '/playground',
